test: preserve CGFloat precision in divider assertions

+++ cmuxTests/BrowserPanelTests.swift
@@ -1739,8 +1739,8 @@ final class WindowBrowserHostViewTests: XCTestCase {
         )
         XCTAssertEqual(
             changedTranslation,
-            32,
-            accuracy: 0.5,
+            CGFloat(32),
+            accuracy: CGFloat(0.5),
             "The forwarded drag must reach the native sidebar tracker"
         )
 
@@ -1796,8 +1796,8 @@ final class WindowBrowserHostViewTests: XCTestCase {
         )
         XCTAssertEqual(
             changedTranslation,
-            28,
-            accuracy: 0.5,
+            CGFloat(28),
+            accuracy: CGFloat(0.5),
             "A reparented Dock divider must continue receiving native drag translation"
         )
     }
